refactor on functions.py, removed need to authenticate on /report command, update lezioni.py to work as esami.py
